Create Customer-queues using linked lists.cpp
This C++ program implements a queue-based request management system using a linked list. It supports the following operations:

Add Request: Adds a new request to the end of the queue.
Process Request: Removes and processes the request at the front of the queue.
View Front Request: Displays the description of the request at the front of the queue without removing it.
Display All Requests: Lists all requests in the queue from front to back.